Change the name of the project to Jupyter Documentation
When we did user testing, a user got to the docs page and tried to click on the "Jupyter" home link at the top left of the page. The user thought it would take him back to the main jupyter.org site. This clarifies that the link doesn't go back to the home page.