Reorganise install page
I went to this page and found there's so much header material that I
have to scroll to see installation instructions even on a desktop
monitor.

This rearranges the document so you get to the useful information
quicker. In particular:

- Moved the Python requirements box down to the 'Experienced Python
user' section. People who will install using Anaconda don't need to
think about it.
- Moved the see also box on installing individual subprojects to the
bottom of the document; it's not important enough to warrant going at
the top.
- Removed a mostly superfluous heading. From a strictly hierarchical
view, the heading makes sense, but for scanning through the document I
don't think it helps.

I've also updated the reference to installing the Python 3.4 version of
Anaconda.