Don't install jupyter module
The 'jupyter' module was moved into the jupyter_core distribution.

As far as I can tell, there's no requirement that a distribution
actually includes any importable modules, so the 'jupyter' distribution
can just be a pure metapackage.

I've updated the version number in anticipation of pushing a new release
with this.

Closes jupyter/jupyter_core#85